style: Detail 페이지 질문 섹션 스타일

+++ src/pages/Detail/Detail.tsx
@@ -1,11 +1,28 @@
 import { useNavigate, useParams } from "react-router-dom";
-import { useState, useEffect } from "react";
+import { useState, useEffect, useCallback } from "react";
 import type { QuestionDataType } from "../../stores/page-store";
 import {
+  HeartIcon,
+  ItemContainer,
+  ItemTypo,
+  QuestionBodySection,
   QuestionTitle,
   QuestionTitleSection,
+  QuestionTopContainer,
   QuestionTypo,
   Root,
+  SaveIcon,
+  ContentTypo,
+  QuestionBottomLeftContainer,
+  QuestionBottomContainer,
+  Typo,
+  QuestionBottomRightContainer,
+  AuthorBox,
+  ViewDateContainer,
+  AskedTypo,
+  AuthorContainer,
+  AuthorProfile,
+  UserStateCircle,
 } from "./styled";
 import DOMPurify from "dompurify";
 import axios from "axios";
@@ -18,8 +35,13 @@ export const Detail = () => {
 
   let { id } = useParams<{ id: string }>();
 
-  const deleteQuestion = async () => {
+  const deleteQuestion = useCallback(async () => {
     try {
+      //삭제할건지 확인
+      if (!window.confirm("정말 삭제하시겠습니까?")) {
+        return;
+      }
+      //삭제
       await axios.delete(`/api/articles/${id}`).then((res) => {
         console.log(res);
         alert("질문 삭제 성공!");
@@ -29,9 +51,9 @@ export const Detail = () => {
       console.error(error);
       alert("질문 삭제 실패!");
     }
-  };
+  }, [id]);
 
-  const fetchData = async () => {
+  const fetchData = useCallback(async () => {
     try {
       const response = await axios.get("/api/articles");
       // id와 일치하는 데이터를 찾아서 setCurrentQuestion에 넣어준다.
@@ -45,7 +67,7 @@ export const Detail = () => {
       console.error(error);
       alert("게시판 정보 가져오기 실패!");
     }
-  };
+  }, [id]);
 
   useEffect(() => {
     // Find the question with the matching ID
@@ -66,21 +88,59 @@ export const Detail = () => {
         <QuestionTypo>Q</QuestionTypo>
         <QuestionTitle>{currentQuestion?.title}</QuestionTitle>
       </QuestionTitleSection>
-      <p>Author: {currentQuestion?.author}</p>
-      <p>Votes: {currentQuestion?.votes}</p>
-      <p>Answers: {currentQuestion?.answers}</p>
-      <p>Views: {currentQuestion?.views}</p>
-      <p>Hashtags: {currentQuestion?.hashtags}</p>
-      <p>Created: {currentQuestion?.createdAt}</p>
-      <p>Updated: {currentQuestion?.updatedAt}</p>
-
-      {/* <p>Content: {currentQuestion?.content}</p> */}
-      <div
-        dangerouslySetInnerHTML={{
-          __html: DOMPurify.sanitize(currentQuestion?.content as string),
-        }}
-      />
-      <button onClick={deleteQuestion}>Delete</button>
+      <QuestionBodySection>
+        <QuestionTopContainer>
+          <ItemContainer>
+            <HeartIcon />
+            <ItemTypo>{currentQuestion?.votes}</ItemTypo>
+            <SaveIcon />
+            <ItemTypo>10</ItemTypo>
+          </ItemContainer>
+          <ItemContainer>
+            <ViewDateContainer>
+              <Typo>조회수 {currentQuestion?.views}</Typo>
+              <Typo>
+                {currentQuestion?.createdAt.substring(
+                  0,
+                  currentQuestion?.createdAt.indexOf("T")
+                )}
+              </Typo>
+            </ViewDateContainer>
+            <ContentTypo
+              dangerouslySetInnerHTML={{
+                __html: DOMPurify.sanitize(currentQuestion?.content as string),
+              }}
+            />
+          </ItemContainer>
+        </QuestionTopContainer>
+        <QuestionBottomContainer>
+          <QuestionBottomLeftContainer>
+            <Typo underline pointer>
+              공유
+            </Typo>
+            <Typo underline pointer>
+              수정
+            </Typo>
+            <Typo underline pointer onClick={deleteQuestion}>
+              삭제
+            </Typo>
+          </QuestionBottomLeftContainer>
+          <QuestionBottomRightContainer>
+            <AuthorBox>
+              <AskedTypo>Asked</AskedTypo>
+              <AuthorContainer>
+                <AuthorProfile>딩동</AuthorProfile>
+                <UserStateCircle />
+                <Typo>15</Typo>
+              </AuthorContainer>
+            </AuthorBox>
+          </QuestionBottomRightContainer>
+        </QuestionBottomContainer>
+        {/* <p>Answers: {currentQuestion?.answers}</p> */}
+        {/*         <p>Hashtags: {currentQuestion?.hashtags}</p>
+        <p>Created: {currentQuestion?.createdAt}</p> */}
+        {/* <p>Updated: {currentQuestion?.updatedAt}</p> */}
+      </QuestionBodySection>
     </Root>
   );
 };
